requirements.txt

Flask==3.0.3
matplotlib==3.7.5
numpy
pillow==10.3.0
redis==5.0.6
torch
torchvision
tensorboard==2.13.0
celery
azure-storage-blob
